+++ Sources/Tokenizer/Elements/PageControl.swift
@@ -20,3 +20,23 @@ public class PageControl: View {
     }
     #endif
 }
+
+public class PageControlProperties: ViewProperties {
+    public let currentPage: AssignablePropertyDescription<Int>
+    public let numberOfPages: AssignablePropertyDescription<Int>
+    public let hidesForSinglePage: AssignablePropertyDescription<Bool>
+    public let pageIndicatorTintColor: AssignablePropertyDescription<UIColorPropertyType>
+    public let currentPageIndicatorTintColor: AssignablePropertyDescription<UIColorPropertyType>
+    public let defersCurrentPageDisplay: AssignablePropertyDescription<Bool>
+    
+    public required init(configuration: Configuration) {
+        currentPage = configuration.property(name: "currentPage")
+        numberOfPages = configuration.property(name: "numberOfPages")
+        hidesForSinglePage = configuration.property(name: "hidesForSinglePage")
+        pageIndicatorTintColor = configuration.property(name: "pageIndicatorTintColor")
+        currentPageIndicatorTintColor = configuration.property(name: "currentPageIndicatorTintColor")
+        defersCurrentPageDisplay = configuration.property(name: "defersCurrentPageDisplay")
+        
+        super.init(configuration: configuration)
+    }
+}
