+++ Sources/Tokenizer/Elements/PickerView.swift
@@ -20,3 +20,10 @@ public class PickerView: View {
     }
     #endif
 }
+
+public class PickerViewProperties: ViewProperties {
+    public required init(configuration: Configuration) {
+        
+        super.init(configuration: configuration)
+    }
+}
