+++ Sources/Tokenizer/Elements/TabBar.swift
@@ -20,3 +20,28 @@ public class TabBar: View {
     }
     #endif
 }
+
+public class TabBarProperties: ViewProperties {
+    public let isTranslucent: AssignablePropertyDescription<Bool>
+    public let barStyle: AssignablePropertyDescription<BarStyle>
+    public let barTintColor: AssignablePropertyDescription<UIColorPropertyType>
+    public let itemSpacing: AssignablePropertyDescription<Float>
+    public let itemWidth: AssignablePropertyDescription<Float>
+    public let backgroundImage: AssignablePropertyDescription<Image>
+    public let shadowImage: AssignablePropertyDescription<Image>
+    public let selectionIndicatorImage: AssignablePropertyDescription<Image>
+    
+    public required init(configuration: Configuration) {
+        isTranslucent = configuration.property(name: "isTranslucent", key: "translucent")
+        barStyle = configuration.property(name: "barStyle")
+        barTintColor = configuration.property(name: "barTintColor")
+        itemSpacing = configuration.property(name: "itemSpacing")
+        itemWidth = configuration.property(name: "itemWidth")
+        backgroundImage = configuration.property(name: "backgroundImage")
+        shadowImage = configuration.property(name: "shadowImage")
+        selectionIndicatorImage = configuration.property(name: "selectionIndicatorImage")
+        
+        super.init(configuration: configuration)
+    }
+}
+
