+++ Sources/Tokenizer/Elements/TextView.swift
@@ -19,3 +19,23 @@ public class TextView: View {
     }
     #endif
 }
+
+public class TextViewProperties: ViewProperties {
+    public let text: AssignablePropertyDescription<TransformedText>
+    public let font: AssignablePropertyDescription<Font>
+    public let textColor: AssignablePropertyDescription<UIColorPropertyType>
+    public let textAlignment: AssignablePropertyDescription<TextAlignment>
+    public let textContainerInset: AssignablePropertyDescription<EdgeInsets>
+    public let allowsEditingTextAttributes: AssignablePropertyDescription<Bool>
+    
+    public required init(configuration: Configuration) {
+        text = configuration.property(name: "text")
+        font = configuration.property(name: "font")
+        textColor = configuration.property(name: "textColor")
+        textAlignment = configuration.property(name: "textAlignment")
+        textContainerInset = configuration.property(name: "textContainerInset")
+        allowsEditingTextAttributes = configuration.property(name: "allowsEditingTextAttributes")
+        
+        super.init(configuration: configuration)
+    }
+}
