+++ Sources/Tokenizer/Elements/Toolbar.swift
@@ -20,3 +20,17 @@ public class Toolbar: View {
     }
     #endif
 }
+
+public class ToolbarProperties: ViewProperties {
+    public let isTranslucent: AssignablePropertyDescription<Bool>
+    public let barStyle: AssignablePropertyDescription<BarStyle>
+    public let barTintColor: AssignablePropertyDescription<UIColorPropertyType>
+    
+    public required init(configuration: Configuration) {
+        isTranslucent = configuration.property(name: "isTranslucent", key: "translucent")
+        barStyle = configuration.property(name: "barStyle")
+        barTintColor = configuration.property(name: "barTintColor")
+        
+        super.init(configuration: configuration)
+    }
+}
